+++ src/cpp/session/modules/SessionDataViewer.R
@@ -313,7 +313,8 @@
 .rs.addFunction("removeCachedData", function(cacheKey, cacheDir)
 {
   # remove data from the cache environment
-  if (exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
+  if (exists(".rs.CachedDataEnv") &&
+      exists(cacheKey, where = .rs.CachedDataEnv, inherits = FALSE))
     rm(list = cacheKey, envir = .rs.CachedDataEnv, inherits = FALSE)
 
   # remove data from the cache directory
@@ -344,14 +345,17 @@
   })
 
   # clean the cache environment
-  rm(list = ls(.rs.CachedDataEnv), where = .rs.CachedDataEnv)
+  # can generate warnings if .rs.CachedDataEnv disappears (we call this on
+  # shutdown); suppress these
+  suppressWarnings(rm(list = ls(.rs.CachedDataEnv), where = .rs.CachedDataEnv))
 
   invisible(NULL)
 })
 
 .rs.addFunction("findWorkingData", function(cacheKey)
 {
-  if (exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
+  if (exists(".rs.WorkingDataEnv") &&
+      exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
     get(cacheKey, envir = .rs.WorkingDataEnv, inherits = FALSE)
   else
     NULL
@@ -359,7 +363,8 @@
 
 .rs.addFunction("removeWorkingData", function(cacheKey)
 {
-  if (exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
+  if (exists(".rs.WorkingDataEnv") &&
+      exists(cacheKey, where = .rs.WorkingDataEnv, inherits = FALSE))
     rm(list = cacheKey, envir = .rs.WorkingDataEnv, inherits = FALSE)
   invisible(NULL)
 })