+++ src/gwt/src/org/rstudio/studio/client/projects/ui/newproject/NewPackagePage.java
@@ -16,11 +16,16 @@ package org.rstudio.studio.client.projects.ui.newproject;
 
 import org.rstudio.core.client.js.JsUtil;
 import org.rstudio.core.client.widget.SelectWidget;
+import org.rstudio.studio.client.projects.Projects;
 import org.rstudio.studio.client.projects.model.NewPackageOptions;
 import org.rstudio.studio.client.projects.model.NewProjectInput;
 
+import com.google.gwt.core.client.Scheduler;
+import com.google.gwt.core.client.Scheduler.ScheduledCommand;
 import com.google.gwt.event.dom.client.ChangeEvent;
 import com.google.gwt.event.dom.client.ChangeHandler;
+import com.google.gwt.event.dom.client.KeyDownEvent;
+import com.google.gwt.event.dom.client.KeyDownHandler;
 import com.google.gwt.user.client.ui.HorizontalPanel;
 
 
@@ -33,6 +38,26 @@ public class NewPackagePage extends NewDirectoryPage
             "Create R Package",
             NewProjectResources.INSTANCE.packageIcon(),
             NewProjectResources.INSTANCE.packageIconLarge());
+      
+      styles_ = NewProjectResources.INSTANCE.styles();
+      
+      txtProjectName_.addKeyDownHandler(new KeyDownHandler()
+      {
+         @Override
+         public void onKeyDown(KeyDownEvent event)
+         {
+            Scheduler.get().scheduleDeferred(new ScheduledCommand()
+            {
+               
+               @Override
+               public void execute()
+               {
+                  validatePackageName();
+               }
+            });
+         }
+      });
+      
    }
    
    protected boolean getOptionsSideBySide()
@@ -89,10 +114,28 @@ public class NewPackagePage extends NewDirectoryPage
    protected NewPackageOptions getNewPackageOptions()
    {
       return NewPackageOptions.create(
+            getProjectName(),
             listProjectType_.getValue().equals("package-rcpp"),  
             JsUtil.toJsArrayString(listCodeFiles_.getCodeFiles()));
    }
    
+   private void validatePackageName()
+   {
+      String packageName = txtProjectName_.getText().trim();
+      
+      // Don't validate if the name is empty
+      if (packageName.isEmpty() || isPackageNameValid(packageName))
+         txtProjectName_.removeStyleName(styles_.invalidPkgName());
+      else
+         txtProjectName_.addStyleName(styles_.invalidPkgName());
+   }
+   
+   private boolean isPackageNameValid(String packageName)
+   {
+      return Projects.PACKAGE_NAME_PATTERN.test(packageName);
+   }
+   
    private SelectWidget listProjectType_;
    private CodeFilesList listCodeFiles_;
+   private final NewProjectResources.Styles styles_;
 }