+++ src/gwt/src/org/rstudio/studio/client/server/remote/ClientEvent.java
@@ -105,8 +105,8 @@ class ClientEvent extends JavaScriptObject
    public static final String RmdTemplateDiscovered = "rmd_template_discovered";
    public static final String RmdTemplateDiscoveryCompleted = "rmd_template_discovery_completed";
    public static final String RmdShinyDocStarted = "rmd_shiny_doc_started";
-   public static final String ShinyAppsDeploymentOutput = "shiny_apps_deployment_output";
-   public static final String ShinyAppsDeploymentCompleted = "shiny_apps_deployment_completed";
+   public static final String RSConnectDeploymentOutput = "rsconnect_deployment_output";
+   public static final String RSConnectDeploymentCompleted = "rsconnect_deployment_completed";
    public static final String UserPrompt = "user_prompt";
    public static final String InstallRtools = "install_r_tools";
    public static final String InstallShiny = "install_shiny";