+++ src/gwt/src/org/rstudio/studio/client/shiny/ui/ShinyApplicationPanel.java
@@ -31,9 +31,9 @@ import org.rstudio.core.client.widget.Toolbar;
 import org.rstudio.core.client.widget.ToolbarButton;
 import org.rstudio.studio.client.workbench.commands.Commands;
 import org.rstudio.studio.client.application.events.EventBus;
+import org.rstudio.studio.client.rsconnect.RSConnect;
+import org.rstudio.studio.client.rsconnect.events.RSConnectActionEvent;
 import org.rstudio.studio.client.shiny.ShinyApplicationPresenter;
-import org.rstudio.studio.client.shiny.ShinyApps;
-import org.rstudio.studio.client.shiny.events.ShinyAppsActionEvent;
 import org.rstudio.studio.client.shiny.model.ShinyApplicationParams;
 
 public class ShinyApplicationPanel extends SatelliteFramePanel<RStudioFrame>
@@ -41,11 +41,11 @@ public class ShinyApplicationPanel extends SatelliteFramePanel<RStudioFrame>
 {
    @Inject
    public ShinyApplicationPanel(Commands commands, EventBus events,
-                                ShinyApps shinyApps)
+                                RSConnect rsconnect)
    {
       super(commands);
       events_ = events;
-      shinyApps.ensureSessionInit();
+      rsconnect.ensureSessionInit();
    }
    
    @Override 
@@ -66,7 +66,7 @@ public class ShinyApplicationPanel extends SatelliteFramePanel<RStudioFrame>
 
       deployButtonSeparator_ = toolbar.addLeftSeparator();
       deployButton_ = new ToolbarButton("Publish", 
-            commands.shinyAppsDeploy().getImageResource(), 
+            commands.rsconnectDeploy().getImageResource(), 
             new ClickHandler()
       {
          @Override
@@ -81,8 +81,8 @@ public class ShinyApplicationPanel extends SatelliteFramePanel<RStudioFrame>
                if (!deployPath.endsWith("/"))
                   deployPath += "/";
                deployPath += "server.R";
-               events_.fireEvent(new ShinyAppsActionEvent(
-                     ShinyAppsActionEvent.ACTION_TYPE_DEPLOY,
+               events_.fireEvent(new RSConnectActionEvent(
+                     RSConnectActionEvent.ACTION_TYPE_DEPLOY,
                      deployPath));
             }
          }