+++ src/gwt/src/org/rstudio/studio/client/workbench/views/console/shell/assist/RCompletionManager.java
@@ -1127,8 +1127,11 @@ public class RCompletionManager implements CompletionManager
          String line)
    {
       int index = line.lastIndexOf('(');
+      String token = line.substring(index + 1);
+      
       AutocompletionContext result = new AutocompletionContext(
-            line.substring(index + 1),
+            token,
+            token,
             AutocompletionContext.TYPE_FILE);
       
       // NOTE: we overload the meaning of the function call string for file
@@ -1225,14 +1228,14 @@ public class RCompletionManager implements CompletionManager
       // trim to cursor position
       firstLine = firstLine.substring(0, input_.getCursorPosition().getColumn());
       
-      // Get the token at the cursor position
-      String token = firstLine.replaceAll(".*[^a-zA-Z0-9._:$@-]", "");
-      
       // If we're in Markdown mode and have an appropriate string, try to get
       // file completions
       if (DocumentMode.isCursorInMarkdownMode(docDisplay_) &&
             firstLine.matches(".*\\[.*\\]\\(.*"))
          return getAutocompletionContextForFileMarkdownLink(firstLine);
+      
+      // Get the token at the cursor position
+      String token = firstLine.replaceAll(".*[^a-zA-Z0-9._:$@-]", "");
       
       // If we're completing an object within a string, assume it's a
       // file-system completion. Note that we may need other contextual information
@@ -1262,8 +1265,9 @@ public class RCompletionManager implements CompletionManager
          return new AutocompletionContext(
                token, AutocompletionContext.TYPE_ROXYGEN);
       
-      // If the token has '$' or '@', escape early as we'll be completing
-      // either from names or an overloaded `$` method
+      // If the token has '$' or '@', add in the autocompletion context --
+      // note that we still need parent contexts to give more information
+      // about the appropriate completion
       if (token.contains("$") || token.contains("@"))
          addAutocompletionContextForDollar(context);
       
@@ -1932,7 +1936,6 @@ public class RCompletionManager implements CompletionManager
          timer_.schedule(400);
       }
       
-      @SuppressWarnings("unused")
       public void cancel()
       {
          timer_.cancel();