+++ NEWS.md
@@ -44,6 +44,9 @@
 
 ### Packages
 
+* Improvements to New Package:
+    - Generate cleaner packages with no warnings
+    - Respect various devtools options
 * Improved checking for supported protocol with packrat package
 * Escape backslashes in library names when loading packages